refactor: split declare-war scoring context-builder into part module (Refs #3749 step 7)
Extract `_DeclareWarTargetContext` (context builder) out of
`diplomatic_candidate_scoring_declare_war.dart` into a new
`diplomatic_candidate_scoring_declare_war_context.dart` part, leaving the
score ladder (entry + suppression chain) in the original file. Pure code
movement within the same `diplomatic_candidate_scoring.dart` library; all
members stay library-private parts and behaviour/determinism are unchanged.

Resulting files: context ~415 LOC, score ladder ~518 LOC (both < 1000).
